# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sanatorio del Salvador internal management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Dashboard summaries and activity
  - name: agenda
    description: Administrative agenda and reminders
  - name: directory
    description: Users and sectors
  - name: auth
    description: Internal user sessions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 6
      responses:
        "200":
          description: Recent activity entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
  /sectors:
    get:
      operationId: listSectors
      tags: [directory]
      summary: List available sectors
      responses:
        "200":
          description: Sectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sector"
  /me:
    get:
      operationId: getCurrentUser
      tags: [directory]
      summary: Get the current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Start an internal user session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: End the current session
      responses:
        "204":
          description: Session ended
  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get the signed-in internal user
      responses:
        "200":
          description: Signed-in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
  /auth/continue-password:
    post:
      operationId: continuePassword
      tags: [auth]
      summary: Keep the temporary password and complete first access
      responses:
        "200":
          description: First access completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordBody"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /agenda:
    get:
      operationId: listAgendaItems
      tags: [agenda]
      summary: List administrative agenda items
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AgendaType"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AgendaStatus"
      responses:
        "200":
          description: Agenda items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgendaItem"
    post:
      operationId: createAgendaItem
      tags: [agenda]
      summary: Create an agenda item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgendaItemInput"
      responses:
        "201":
          description: Created agenda item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgendaItem"
  /agenda/{id}:
    patch:
      operationId: updateAgendaItem
      tags: [agenda]
      summary: Update an agenda item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgendaItemUpdate"
      responses:
        "200":
          description: Updated agenda item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgendaItem"
        "404":
          description: Agenda item not found
    delete:
      operationId: deleteAgendaItem
      tags: [agenda]
      summary: Delete an agenda item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Agenda item not found
  /users:
    get:
      operationId: listUsers
      tags: [directory]
      summary: List internal users for the superadmin
      responses:
        "200":
          description: Internal users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserAdmin"
    post:
      operationId: createUser
      tags: [directory]
      summary: Create an internal user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAdmin"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [directory]
      summary: Update an internal user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAdmin"
    delete:
      operationId: deleteUser
      tags: [directory]
      summary: Delete an internal user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [directory]
      summary: Reset an internal user's password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetUserPasswordBody"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AgendaType:
      type: string
      enum: [cheque, vencimiento, tarea, pago, otro]
    AgendaStatus:
      type: string
      enum: [pendiente, completado]
    AgendaItem:
      type: object
      required: [id, title, dueDate, type, status, responsibleName, sectorName, description, reminderDate]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date
        type:
          $ref: "#/components/schemas/AgendaType"
        status:
          $ref: "#/components/schemas/AgendaStatus"
        responsibleName:
          type: string
        sectorName:
          type: ["string", "null"]
        reminderDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
    AgendaItemInput:
      type: object
      required: [title, dueDate, type, responsibleName]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        dueDate:
          type: string
          format: date
        type:
          $ref: "#/components/schemas/AgendaType"
        responsibleName:
          type: string
          minLength: 1
        sectorName:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/AgendaStatus"
    AgendaItemUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        dueDate:
          type: string
          format: date
        type:
          $ref: "#/components/schemas/AgendaType"
        responsibleName:
          type: string
          minLength: 1
        sectorName:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/AgendaStatus"
    Sector:
      type: object
      required: [id, name, shortName]
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        members:
          type: integer
    CurrentUser:
      type: object
      required: [id, username, name, email, role, active, mustChangePassword, modules, sectors]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [superadmin, responsable, usuario]
        active:
          type: boolean
        mustChangePassword:
          type: boolean
        modules:
          type: array
          items:
            type: string
        sectors:
          type: array
          items:
            $ref: "#/components/schemas/Sector"
    AuthUser:
      allOf:
        - $ref: "#/components/schemas/CurrentUser"
        - type: object
          required: [sectorIds]
          properties:
            sectorIds:
              type: array
              items:
                type: integer
    LoginBody:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    LoginResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
    UpdatePasswordBody:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string
          minLength: 6
    ActionResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    UserAdmin:
      allOf:
        - $ref: "#/components/schemas/AuthUser"
        - type: object
          required: [createdAt, lastLoginAt]
          properties:
            createdAt:
              type: string
              format: date-time
            lastLoginAt:
              type: ["string", "null"]
              format: date-time
    CreateUserBody:
      type: object
      required: [username, name, password, role]
      properties:
        username:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        email:
          type: ["string", "null"]
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [responsable, usuario]
        active:
          type: boolean
        modules:
          type: array
          items:
            type: string
        sectorIds:
          type: array
          items:
            type: integer
    UpdateUserBody:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [responsable, usuario]
        active:
          type: boolean
        modules:
          type: array
          items:
            type: string
        sectorIds:
          type: array
          items:
            type: integer
    ResetUserPasswordBody:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 6
    DashboardSummary:
      type: object
      required: [upcomingReminders, pendingGuardias, lowStockItems, recentActivityCount]
      properties:
        upcomingReminders:
          type: integer
        pendingGuardias:
          type: integer
        lowStockItems:
          type: integer
        recentActivityCount:
          type: integer
        nextReminder:
          type: ["string", "null"]
        nextReminderDate:
          type: ["string", "null"]
          format: date
    Activity:
      type: object
      required: [id, title, detail, actor, createdAt, kind]
      properties:
        id:
          type: integer
        title:
          type: string
        detail:
          type: string
        actor:
          type: string
        createdAt:
          type: string
          format: date-time
        kind:
          type: string
          enum: [agenda, guardia, inventario, instructivo, sistema]