# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sanatorio del Salvador internal management API
 * OpenAPI spec version: 0.1.0
 */
import type { CurrentUser } from './currentUser';

export type AuthUser = CurrentUser & {
  sectorIds: number[];
};