# lib/api-zod/src/generated/types/createUserBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sanatorio del Salvador internal management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateUserBodyRole } from './createUserBodyRole';

export interface CreateUserBody {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  name: string;
  /** @nullable */
  email?: string | null;
  /** @minLength 6 */
  password: string;
  role: CreateUserBodyRole;
  active?: boolean;
  modules?: string[];
  sectorIds?: number[];
}