# lib/api-zod/src/generated/types/currentUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sanatorio del Salvador internal management API
 * OpenAPI spec version: 0.1.0
 */
import type { CurrentUserRole } from './currentUserRole';
import type { Sector } from './sector';

export interface CurrentUser {
  id: number;
  username: string;
  name: string;
  /** @nullable */
  email: string | null;
  role: CurrentUserRole;
  active: boolean;
  mustChangePassword: boolean;
  modules: string[];
  sectors: Sector[];
}