# lib/api-zod/src/generated/types/loginBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sanatorio del Salvador internal management API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginBody {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}