# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sanatorio del Salvador internal management API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface LoginResponse {
  user: AuthUser;
}