# lib/api-zod/src/generated/types/resetUserPasswordBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sanatorio del Salvador internal management API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetUserPasswordBody {
  /** @minLength 6 */
  password: string;
}