# lib/api-zod/src/generated/types/updatePasswordBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sanatorio del Salvador internal management API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePasswordBody {
  /** @minLength 6 */
  newPassword: string;
}