# lib/api-zod/src/generated/types/userAdmin.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sanatorio del Salvador internal management API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export type UserAdmin = AuthUser & ({
  createdAt: Date;
  /** @nullable */
  lastLoginAt: Date | null;
});